Use a client authentication supplier rather than a static client authentication
This allows authenticating using JWT, where a client authentication
instance needs to be created for each authenticated request.

This is a breaking change, but the Configuration class still has
a backwards-compatible constructor, and users aren't expected to
actually make use of the instance after construction, so this
shouldn't actually break projects in practice (and we're still in
RC so such a breaking change is acceptable as the API matures).

Fixes #52