Remove OIDC-Servlets+RS AuthenticationRedirector subclass
This will allow easier subclassing of the base class for
specific usage, such as JAR and PAR, and possibly DPoP.

The class was used both as a "strategy" to be subclassed
for customization, and a public API for others to call
(such as the authorization filters), but the JAX-RS-specific
subclass changing the API (to adapt it to JAX-RS) made the
class responsibility blurrier.

The JAX-RS specific API was moved to static methods in the
Utils class (now public). The base class keeps its public API